refactor: move type product to a distinct file
