+++ src/data-loader.ts
@@ -1,16 +1,4 @@
-type product = {
-  id: number;
-  title: string;
-  description: string;
-  price: number;
-  discountPercentage: number;
-  rating: number;
-  stock: number;
-  brand: string;
-  category: string;
-  thumbnail: string;
-  images: string[];
-};
+import product from "./interfaces/product";
 
 type responseData = {
   products: product[];
